new changes applied

+++ firebase.js
@@ -1,5 +1,3 @@
-
-
   $( document ).ready( function() {
     const firebaseConfig = {
         apiKey: "AIzaSyDlsoAqDJA213Z40hLcSLcDOYuHeMphUWE",
@@ -11,6 +9,76 @@
         appId: "1:75419734917:web:fc352a3abda9bc74506aec"
       };
       firebase.initializeApp(firebaseConfig);
+var day=10; 
+var month=5;
+var year=2020; 
+var today = new Date();
+var currentMonth = today.getMonth();
+var currentYear = today.getFullYear();
+var selectYear = document.getElementById("year");
+var selectMonth = document.getElementById("month");
+console.log(currentMonth, currentYear);
+
+var months = ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"];
+
+var monthAndYear = document.getElementById("monthAndYear");
+function daysInMonth(iMonth, iYear) {
+  return 32 - new Date(iYear, iMonth, 32).getDate();
+}
+function show(month, year,day) {
+
+  let firstDay = (new Date(year, month)).getDay();
+
+  var tbl = document.getElementById("calendar-body"); // body of the calendar
+
+  // clearing all previous cells
+  tbl.innerHTML = "";
+
+  // filing data about month and in the page via DOM.
+  monthAndYear.innerHTML = months[month] + " " + year;
+  selectYear.value = year;
+  selectMonth.value = month;
+
+  // creating all cells
+  let date = 1;
+  for (let i = 0; i < 6; i++) {
+      // creates a table row
+      let row = document.createElement("tr");
+
+      //creating individual cells, filing them up with data.
+      for (let j = 0; j < 7; j++) {
+          if (i === 0 && j < firstDay) {
+              var cell = document.createElement("td");
+              var cellText = document.createTextNode("");
+              cell.appendChild(cellText);
+              row.appendChild(cell);
+          }
+          else if (date > daysInMonth(month, year)) {
+              break;
+          }
+
+          else {
+              var cell = document.createElement("td");
+              var cellText = document.createTextNode(date);
+              if (date === today.getDate() && year === today.getFullYear() && month === today.getMonth()) {
+                  cell.classList.add("bg-info");
+              } // color today's date
+              cell.appendChild(cellText);
+              row.appendChild(cell);
+              if (day===date){
+                cell.style.color="white";
+                cell.style.fontWeight="bold";
+                cell.bgColor = '#43a5fc';
+              }
+              date++;
+          }
+
+
+      }
+
+      tbl.appendChild(row); // appending each row into calendar body.
+  }
+}
 var db = firebase.firestore();
   // Your web app's Firebase configuration
   var appointments = db.collection('appointments');
@@ -19,8 +87,10 @@ var db = firebase.firestore();
         if (!doc.exists){
             document.getElementById("mytable").deleteRow(0);
             table();
+            show(month,year,0);
         }
         else{
+        show(month,year,day);
         data = doc.data();
         var name=data["accountName"];
         var therapist=data["therapistName"];
@@ -43,11 +113,22 @@ var db = firebase.firestore();
       var newCell3 = newRow.insertCell(2);
       newCell2.innerHTML="No upcoming visits";
       }}
-     table();
-  $('body').on('click', 'input.deleteDep', function() {   
-    $(this).closest('tr').remove();
-    alert( "You have canceled the tharapy session" );
-    appointments.doc('Mia->Karen').delete();
-    table();
-                                              })
+  table();
+
+    function myFunction() {
+      var modal = document.getElementById('try');
+      modal.onclick= function() {
+      var txt;
+      var r = confirm("Are you sure to delete your appointment?");
+      if (r == true) {
+        show(month,year,0);
+        $(this).closest('tr').remove();
+        table();
+        appointments.doc('Mia->Karen').delete();
+      } else {
+      }
+    }                                 
+  }
+  myFunction();
+
   })
